fix: Share link of the OTQ result is changed when selecting new properties in query form
The problem is related to angular automatic change detection mechanism and the leaflet library registering e.g. mouseover events. This leads to component updates on every mousemove event and triggering the share link function to get the pemalink url again and again, but it should only happen once on component creation.
The fix is to detach the oqt result component from angulars change detection such that it don't get updates from the input form once its created.
That was already implemented in the ohsome-result but was missing in the oqt-result-component

Closes: #23