fix: attribute completeness dropdown search bar initialization and styling
Previously the search bar was not initialized properly when the attribute-completeness indicator was unselected during component initialization. Then search functionality was not available.
Furthermore the height of the text input field for the search was not set propertly because it inherited wrong rules from SemanicCSS's .ui.toggle.checkbox input which is actually meant for input[type=checkbox] but matched out input[type=text] from the search dropdown.

Refs: #42